+++ miamoremustachio/ToDo/modules/middlewares/validation/subtask-validation.js
@@ -1,21 +1,6 @@
-const { ValidationError } = require('../../errors/validation-error.js');
-const { TITLE_LENGTH } = require('../../helpers/constants.js');
+const { checkSubtask } = require('../../validators/subtask-validator.js');
 
-const { MIN, MAX } = TITLE_LENGTH;
-
-const checkSubtask = {
-  entity: 'Subtask',
-  messages: {
-    title: `Invalid title (only strings between ${MIN} and ${MAX} characters are allowed).`,
-  },
-  title(title) {
-    if (title.length < MIN || title.length > MAX) {
-      throw new ValidationError(this.messages.title, this.entity);
-    }
-  },
-};
-
-const checkSubtaskFields = (req, res, next) => {
+const validateSubtask = (req, res, next) => {
   const { title } = req.body;
 
   if (title) {
@@ -30,4 +15,4 @@ const checkSubtaskFields = (req, res, next) => {
   next();
 }
 
-module.exports = { checkSubtaskFields };
+module.exports = { validateSubtask };
