+++ miamoremustachio/ToDo/modules/validators/subtask-validator.js
@@ -1,0 +1,18 @@
+const { ValidationError } = require('../errors/validation-error.js');
+const { TITLE_LENGTH } = require('../helpers/constants.js');
+
+const { MIN, MAX } = TITLE_LENGTH;
+
+const checkSubtask = {
+  entity: 'Subtask',
+  messages: {
+    title: `Invalid title (only strings between ${MIN} and ${MAX} characters are allowed).`,
+  },
+  title(title) {
+    if (title.length < MIN || title.length > MAX) {
+      throw new ValidationError(this.messages.title, this.entity);
+    }
+  },
+};
+
+module.exports = { checkSubtask };
